Fixed iOS framework identifier generation and bump `archive` dependency (#185)
* Fix leading dash in framework identifier generation

Ensures that any leading dashes are removed from the generated framework identifier in xcframework_utils.sh. Defaults to 'framework' if the identifier is empty after processing.

* Update CFBundlePackageType to FMWK in plist generation

Changed the CFBundlePackageType from 'APPL' to 'FMWK' in the create_plist function to correctly identify the bundle as a framework. Also made a minor whitespace adjustment in create_xcframework_from_dylibs.

* Bump `archive` dependency to `^4.0.7` (#184)

* update `archive` dependency to ``^4.0.7` and bump serious_python version to `0.9.5`

* revert flutter_lints dependency to version 2.0.0

* Bump version to 0.9.5 for Android and Darwin

Updated the version number to 0.9.5 in both the Android build.gradle and Darwin podspec files to prepare for a new release.

---------

Co-authored-by: Feodor Fitsner <feodor@appveyor.com>

* Update changelogs for iOS framework identifier fix

Added notes to all package changelogs for version 0.9.5 about the fix for iOS framework identifier generation.

---------

Co-authored-by: TheEthicalBoy <98978078+ndonkoHenri@users.noreply.github.com>