Merge Android crash fix into the main branch (#7)
* Fixed iOS framework identifier generation and bump `archive` dependency (#185)

* Fix leading dash in framework identifier generation

Ensures that any leading dashes are removed from the generated framework identifier in xcframework_utils.sh. Defaults to 'framework' if the identifier is empty after processing.

* Update CFBundlePackageType to FMWK in plist generation

Changed the CFBundlePackageType from 'APPL' to 'FMWK' in the create_plist function to correctly identify the bundle as a framework. Also made a minor whitespace adjustment in create_xcframework_from_dylibs.

* Bump `archive` dependency to `^4.0.7` (#184)

* update `archive` dependency to ``^4.0.7` and bump serious_python version to `0.9.5`

* revert flutter_lints dependency to version 2.0.0

* Bump version to 0.9.5 for Android and Darwin

Updated the version number to 0.9.5 in both the Android build.gradle and Darwin podspec files to prepare for a new release.

---------

Co-authored-by: Feodor Fitsner <feodor@appveyor.com>

* Update changelogs for iOS framework identifier fix

Added notes to all package changelogs for version 0.9.5 about the fix for iOS framework identifier generation.

---------

Co-authored-by: TheEthicalBoy <98978078+ndonkoHenri@users.noreply.github.com>

* Fix zip directory (#186)

* Make zipDirectory call asynchronous

Added 'await' to the zipDirectory method to ensure the archive creation completes before proceeding. This change improves reliability when creating app archives.

* Bump to 0.9.6 and make zipDirectory asynchronous

Updated all platform packages to version 0.9.6. The zipDirectory call is now asynchronous across all implementations for improved performance and consistency.

* Make setenv calls awaitable in Python init

Refactored the setenv function to return a Future<void> and updated all calls to setenv to use await. This ensures environment variables are set in the correct order before proceeding.

* Ensure single CPython interpreter per process

Introduces a guard to prevent repeated initialization and finalization of the CPython interpreter, addressing crashes on second launch with CPython 3.12. Also ensures the current native thread is registered with the GIL when running Python code, improving stability when re-entering Python from new Dart isolates or threads.

* Remove 'Warning:' prefix from debug message

Updated the debugPrint statement to remove the 'Warning:' prefix when unable to load libpyjni.so. This change streamlines log output for consistency.

* Revert "Ensure single CPython interpreter per process"

This reverts commit 3f4adc08b82147a46452b19d2a29d6af715a58a0.

* chore: bump python version to 3.14.2

* fix python-windows-dart package name

---------

Co-authored-by: Feodor Fitsner <feodor@appveyor.com>
Co-authored-by: TheEthicalBoy <98978078+ndonkoHenri@users.noreply.github.com>